added GDPR endpoint

+++ src/main/java/com/semester6/UserRegistrationMicroService/repositories/IAccountRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 
-public interface IAccountRepository extends JpaRepository<Account, Integer>
+public interface IAccountRepository extends JpaRepository<Account, Long>
 {
     @Query(value = "SELECT * FROM `account` ORDER BY `id` DESC LIMIT 1", nativeQuery = true)
     public Account findLastCreatedAccount();
